+++ instrumentation/opentelemetry-instrumentation-pyramid/tests/pyramid_base_test.py
@@ -24,6 +24,8 @@ class InstrumentationTest:
         helloid = int(request.matchdict["helloid"])
         if helloid == 500:
             raise exc.HTTPInternalServerError()
+        if helloid == 404:
+            raise exc.HTTPNotFound()
         if helloid == 302:
             raise exc.HTTPFound()
         if helloid == 204: